fix migration

+++ database/migrations/2025_08_08_074518_remove_tour_time_id_from_tours_table.php
@@ -9,8 +9,9 @@ return new class extends Migration
     public function up()
     {
         Schema::table('tours', function (Blueprint $table) {
-            $table->dropForeign(['tour_time_id']);
-            $table->dropColumn('tour_time_id');
+            if (Schema::hasColumn('tours', 'tour_time_id')) {
+                $table->dropColumn('tour_time_id');
+            }
         });
     }
